Move overload test to separate class, rearrange test dir
