+++ tests/JustGivingClientTest.php
@@ -86,38 +86,4 @@ class JustGivingClientTest extends TestCase
 
         $client->Account->retrieve();
     }
-
-    /**
-     * Get a mock AuthValue object.
-     *
-     * @return AuthValue|\Mockery\MockInterface
-     */
-    private function getAuthMock()
-    {
-        $auth = Mockery::mock(AuthValue::class);
-        $auth->shouldReceive('getHeaders')
-            ->andReturn([]);
-
-        return $auth;
-    }
-
-    /**
-     * @runInSeparateProcess as we're using an overload mock, which interferes with the autoloading process
-     * @test
-     */
-    public function it_automatically_instantiates_a_psr_18_client()
-    {
-        // Overload the Guzzle client class to make sure the JustGiving client is instantiating it internally
-        $http = Mockery::mock('overload:' . Client::class);
-        $http->shouldReceive('send')
-            ->withArgs(function (RequestInterface $request) {
-                return $request->getUri()->__toString() === 'https://api.justgiving.com/v1/account';
-            })
-            ->once()
-            ->andReturn(new Response);
-
-        $client = new JustGivingClient($this->getAuthMock());
-
-        $client->Account->retrieve();
-    }
 }
